+++ CHANGELOG.md
@@ -13,6 +13,10 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/) and this p
 options to replace the graph or add the result shapes to the graph, or stop the workflow if the specified output graph 
 already exists.
 
+### Added
+
+- Add dcterms:created datetime for new/replaced output graphs or dcterms:modified datetime for updated output graphs
+
 
 ## [1.0.0] 2025-02-03